Fix relative path in Xcode project
### The problem:
`libz.1.dylib` is linked as a relative path (`../../../../../libz.1.dylib`). It works if the project is not too deep in the file structure. If it is deep enough, the amount of `..` doesn't suffice.

### The solution:
Use absolute path for the dylib, so it works no matter how deep in the hierarchy the project is.